requirements.txt
streamlit>=1.24.0
python-dotenv>=0.19.0
google-cloud-aiplatform>=1.25.0
pandas>=1.5.0
plotly>=5.13.0
pdfkit>=1.0.0
python-docx>=0.8.11
numpy>=1.21.0
email-validator>=1.1.3
matplotlib>=3.8.3
Pillow>=10.2.0
PyMuPDF>=1.23.26
tiktoken>=0.6.0
requests>=2.31.0
scikit-learn>=1.3.0
scipy>=1.11.0
openpyxl>=3.1.0
seaborn>=0.12.0